Full Working Sql and Loader
